spun Questions for candidates out into its own component called Agenda;
