+++ app/components/candidate-preamble/index.js
@@ -1,9 +1,10 @@
 'use strict'
 import React, { useState } from 'react'
 import { createUseStyles } from 'react-jss'
-import Button from './button'
+import Button from '../button'
 import cx from 'classnames'
-import ConversationHeader from './conversation-header'
+import ConversationHeader from '../conversation-header'
+import Agenda from './Agenda.js'
 
 /**
  * 
@@ -34,31 +35,6 @@ candidate_questions=[
 function CandidatePreamble({ onClick, agreed, bp_info, subject, candidate_questions, instructionLink, timeLimits }) {
   const classes = useStyles()
   const [isPortrait, togglePortrait] = useState(false)
-  const makeQuestions = (className, questions, keyIndex = 'mq') => {
-    return (
-      <ul className={className}>
-        {questions.map((question, index) =>
-          typeof question === 'string' ? (
-            <li key={keyIndex + '-' + index}>{question}</li>
-          ) : question.length === 1 ? (
-            <li key={keyIndex + '-' + index}>
-              <strong>[ {timeLimits[index]} seconds ]</strong> {question[0]}
-            </li>
-          ) : question[0][0] >= '0' && question[0][0] <= '9' ? (
-            <li>
-              <strong>[ {timeLimits[index]} seconds ]</strong>
-              {makeQuestions(classes.questionListInnerHeadless, question, keyIndex + index)}
-            </li>
-          ) : (
-            <li>
-              <strong>[ {timeLimits[index]} seconds ]</strong> {question[0]}
-              {makeQuestions(classes.questionListInner, question.slice(1), keyIndex + index)}
-            </li>
-          )
-        )}
-      </ul>
-    )
-  }
 
   return (
     <div className={cx(classes['Preamble'], agreed && classes['agreed'])}>
@@ -77,8 +53,8 @@ function CandidatePreamble({ onClick, agreed, bp_info, subject, candidate_questi
           )}
         </h2>
         <p>
-          Ballotpedia and EnCiv are teaming up to create a better way for candidates to be heard, and voters to learn
-          about their candidates.
+          Undebate and EnCiv are teaming up to create a better way for candidates to be heard, and voters to learn about
+          their candidates.
         </p>
         <p>
           You are invited to engage in an application that will include you, as part of a publicly available online
@@ -112,11 +88,7 @@ function CandidatePreamble({ onClick, agreed, bp_info, subject, candidate_questi
             </li>
           )}
         </ul>
-        <h2 style={{ marginBottom: '0.5rem' }}>Questions for Candidates</h2>
-        {makeQuestions(
-          classes.questionList,
-          candidate_questions && candidate_questions.slice(0, -1)
-        ) /* the last thing in the list is the moderators closing remarks*/}
+        <Agenda candidate_questions={candidate_questions} timeLimits={timeLimits} />
         <div className={classes['center']}>
           <Button onClick={onClick}>Next</Button>
         </div>
@@ -162,34 +134,6 @@ const useStyles = createUseStyles({
   },
   portrait: {
     marginTop: '20vh',
-  },
-  questionList: {
-    listStyleType: 'upper-alpha',
-    marginTop: 0,
-    marginBottom: 0,
-    padding: '0',
-    paddingLeft: '2em',
-    '& li': {
-      paddingTop: '0.4em',
-    },
-  },
-  questionListInner: {
-    listStyleType: 'none',
-    paddingLeft: '0',
-    '& li:last-child': {
-      paddingBottom: '0',
-    },
-  },
-  questionListInnerHeadless: {
-    listStyleType: 'none',
-    paddingLeft: '0',
-    paddingTop: '0',
-    '& li:first-child': {
-      paddingTop: '0',
-    },
-    '& li:last-child': {
-      paddingBottom: '0',
-    },
   },
 })
